my_agent_project/src/agent.py
import google.generativeai as genai # Using the placeholder library

def run_agent() -> str:
    """
    A simple agent that returns a greeting.
    """
    # In a real scenario, you might configure an API key here
    # genai.configure(api_key="YOUR_API_KEY")
    # For this example, we'll just return a string
    # without making an API call.

    greeting: str = "Hello, World! This is your agent speaking."
    return greeting

if __name__ == "__main__":
    message: str = run_agent()
    print(message)
